// types/dashboard.d.ts
export type TDashboardCardData = {
    iconClass: string;
    count: string;
    title: string;
    progressTitle: string;
    value: string;
    percentage: string;
    progressBarClass: string;
    progressWidth: string;
}

export type TCardListOneData = {
    iconClass: string;
    title: string;
    status: string;
    mainCount: string;
    chartId: string;
    percentageChange: string;
    changeColorClass: string;
    description: string;
}